+++ NMS/v1_7_R1/src/main/java/com/gmail/filoghost/holographicdisplays/nms/v1_7_R1/EntityNMSSlime.java
@@ -162,8 +162,8 @@ public class EntityNMSSlime extends EntitySlime implements NMSSlime {
 		Entity entity = (Entity) vehicleBase;
 
 		try {
-			ReflectionUtils.setPrivateField(Entity.class, this, "g", (double) 0.0);
-			ReflectionUtils.setPrivateField(Entity.class, this, "h", (double) 0.0);
+			ReflectionUtils.setPrivateField(Entity.class, this, "g", 0.0);
+			ReflectionUtils.setPrivateField(Entity.class, this, "h", 0.0);
 		} catch (Exception ex) {
 			ConsoleLogger.error(ex);
 		}
