+++ NMS/v1_7_R4/src/main/java/com/gmail/filoghost/holographicdisplays/nms/v1_7_R4/EntityNMSHorse.java
@@ -166,8 +166,8 @@ public class EntityNMSHorse extends EntityHorse implements NMSHorse {
 		Entity entity = (Entity) vehicleBase;
 
 		try {
-			ReflectionUtils.setPrivateField(Entity.class, this, "g", (double) 0.0);
-			ReflectionUtils.setPrivateField(Entity.class, this, "h", (double) 0.0);
+			ReflectionUtils.setPrivateField(Entity.class, this, "g", 0.0);
+			ReflectionUtils.setPrivateField(Entity.class, this, "h", 0.0);
 		} catch (Exception ex) {
 			ConsoleLogger.error(ex);
 		}
