+++ NMS/v1_8_R3/src/main/java/com/gmail/filoghost/holographicdisplays/nms/v1_8_R3/EntityNMSSlime.java
@@ -166,8 +166,8 @@ public class EntityNMSSlime extends EntitySlime implements NMSSlime {
 		Entity entity = (Entity) vehicleBase;
 
 		try {
-			ReflectionUtils.setPrivateField(Entity.class, this, "ar", (double) 0.0);
-			ReflectionUtils.setPrivateField(Entity.class, this, "as", (double) 0.0);
+			ReflectionUtils.setPrivateField(Entity.class, this, "ar", 0.0);
+			ReflectionUtils.setPrivateField(Entity.class, this, "as", 0.0);
 		} catch (Exception ex) {
 			ConsoleLogger.error(ex);
 		}
